Include a file system tree view of the diagrams directory
This adds a third tab in the left panel. Users may know
open a graph by its path, instead of looking in the
database. This way, we limit the content of the database to
the exported symbols, and if a more specific graph is looked for, it can
be found in the tree view.